rename s3_inputs
