+++ dev/s3_helper_test.py
@@ -25,7 +25,7 @@ from docling_jobkit.connectors.s3_helper import (
     get_source_files,
 )
 from docling_jobkit.datamodel.convert import ConvertDocumentsOptions
-from docling_jobkit.datamodel.s3_inputs import S3Coordinates
+from docling_jobkit.datamodel.s3_coords import S3Coordinates
 
 
 class Settings(BaseSettings):
