+++ docling_jobkit/datamodel/s3_coords.py
